pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "dragonframe-bg-switcher"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "Markus Vogt" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "typer-slim",
  "flask",
  "turbo-flask",
  "rich",
]

[project.urls]
Documentation = "https://github.com/unknown/dragonframe-bg-switcher#readme"
Issues = "https://github.com/unknown/dragonframe-bg-switcher/issues"
Source = "https://github.com/unknown/dragonframe-bg-switcher"

[project.scripts]
dragonframe-bg-switcher = "dragonframe_bg_switcher.switcher:cli_entrypoint"

[tool.hatch.version]
source = "vcs"

[tool.hatch.envs.default]
dependencies = [
  "coverage[toml]>=6.5",
  "pytest",
]
[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = [
  "- coverage combine",
  "coverage report",
]
cov = [
  "test-cov",
  "cov-report",
]

[[tool.hatch.envs.all.matrix]]
python = ["3.8", "3.9", "3.10", "3.11", "3.12"]

[tool.hatch.envs.types]
dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/dragonframe_bg_switcher tests}"

[tool.coverage.run]
source_pkgs = ["dragonframe_bg_switcher", "tests"]
branch = true
parallel = true
omit = [
  "src/dragonframe_bg_switcher/__about__.py",
]

[tool.coverage.paths]
dragonframe_bg_switcher = ["src/dragonframe_bg_switcher", "*/dragonframe-bg-switcher/src/dragonframe_bg_switcher"]
tests = ["tests", "*/dragonframe-bg-switcher/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
